+reek
+reek

+++ LDB_4lab/app/models/work_group_manager.rb
@@ -6,22 +6,33 @@ require 'yaml'
 
 # Saves and writes work group related data
 class WorkGroupManager
+  def initialize
+    @state = true
+  end
+
+  def stater(var = @state)
+    @state = var
+  end
+
   def save_group(name)
     WorkGroup.create(name: name)
-    true
+    @state
   end
 
   def delete_group(group)
+    return false unless @state
+
     wg = WorkGroup.find_by(id: group)
     wg.destroy
     true
   end
 
   def list_groups
+    return false unless @state
+
     arr = []
-    lofids = WorkGroup.all.ids
-    lofids.each do |t|
-      arr.push(t.to_s + ':' + (WorkGroup.find_by id: t).name)
+    WorkGroup.all.ids.each do |el|
+      arr.push(el.to_s + ':' + (WorkGroup.find_by id: el).name)
     end
 
     arr
